requirements.txt
langchain
python-dotenv
supabase 
streamlit
openai
psycopg2
PyPDF2
pandas
tiktoken
mailersend
pdfreader